modified render library
