+++ src/js/render_utils.js
@@ -1,3 +1,4 @@
+import { createElement } from './createElement';
 import { renderFilter } from './filter';
 import { renderPopular } from './render_popular';
 import { renderSearch } from './render_search';
@@ -30,4 +31,20 @@ export function choiceMainRender() {
     if (getFromStorage('libraryState') === 'Watched') renderWatched();
     if (getFromStorage('libraryState') === 'Queue') renderQueue();
   }
+}
+
+export function addImgNodata() {
+  const img = createElement(
+      'div',
+      {
+        class:`${getFromStorage('theme') === "dark" ? 'nodata-image dark' : 'nodata-image light'}`
+      },
+      ''
+    );
+    document.querySelector('.cards__list').before(img);
+}
+
+export function removeImgNodata() {
+  if(document.querySelector('.cards__list').previousElementSibling)
+    document.querySelector('.cards__list').previousElementSibling.remove();
 }
